GNU General Public License v2
In the original repository (before the fork) there is a LICENSE file for Apache Licence 2.0. However, all source code files and the README file mention that the code is licensed under the GNU General Public License version 2. As these files were added after the LICENSE file by the original author the assumption is that the code is licensed under GNU General Public License version 2.